Add DeepKalmanFilter model
This is the first iteration of the implementation of Deep Kalman Filter. It uses the simplest training procedure possible and the RNN is a simple forward network (I have not implemented the backward RNN yet). The trainig exhibits high variance.

+++ neurolib/models/predictor_rnn.py
@@ -185,7 +185,7 @@ class PredictorRNN(Model):
     else:
       self._check_custom_build()
       builder.scope = self._main_scope
-    data_type = 'int32' if self.is_categorical else 'float32'
+    data_type = 'int32' if self.is_categorical else 'float64'
     is2 = builder.addInputSequence(self.input_dims, name='outputSeq', dtype=data_type)
     os2 = builder.addOutputSequence(name='observation')
     builder.addDirectedLink(is2, os2)
